requirements.txt
# HTTP & Feed
httpx==0.27.0
feedparser==6.0.11
PyYAML==6.0.2

# Content extraction
goose3==3.1.19
beautifulsoup4==4.12.3
# trafilatura==1.9.0
# lxml>=4.9.4,<5.2.0   # ✅ trafilatura 호환 범위


# Database
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9

# OpenAI (sentiment-worker 용)
openai>=1.40.0

# 기타
python-dateutil==2.9.0.post0
prometheus_client==0.20.0

pandas==1.5.3
